requirements.txt
certifi
MonsterLab
pandas
flask
python-dotenv
pymongo
altair
joblib
scikit-learn
